niflheim/i3status: Show memory available if >90% are used.
Until today never noticed this minor error in my configuration.

Note for those not familiar with the difference:

On Linux free memory is completely unused memory, a total waste of this
ressource. Important is the memory categorized as available, this covers
caches/buffers meaning data which isn't required to be kept in-memory
but there for performance and other improvements. Free memory is part of
available memory so if there are 300MiB free, but 25GiB available you
won't have a problem starting the greedy web-browser due to a memory
shortage.

Details can be found in /proc/meminfo and related Linux kernel
documentation. In the latter probably also somewhere the criteria by
which is chosen what to remove from the cache for a process allocation.